corrected solution v2
