+++ src/components/FindMovie/FindMovie.tsx
@@ -1,40 +1,97 @@
-import React from 'react';
+import React, { useCallback, useState } from 'react';
 import classNames from 'classnames';
 import './FindMovie.scss';
 import { Movie } from '../../types/Movie';
 import { MovieCard } from '../MovieCard';
+import { MovieData } from '../../types/MovieData';
+import { ResponseError } from '../../types/ReponseError';
+import { getMovie } from '../../api';
 
 type Props = {
-  query: string,
-  onChangeQuery: (event: React.ChangeEvent<HTMLInputElement>) => void,
-  onSubmit: (query: string) => void,
-  noMovieFound: boolean,
-  onAddToList: () => void,
-  previevMovie: Movie | null,
-  onLoading: () => void,
-  isLoading: boolean,
-  badRequest: boolean,
+  movies: Movie[];
+  onAddToList: (movie: Movie) => void;
 };
 
 export const FindMovie: React.FC<Props> = ({
-  query,
-  onChangeQuery,
-  onSubmit,
-  noMovieFound,
+  movies,
   onAddToList,
-  previevMovie,
-  onLoading,
-  isLoading,
-  badRequest,
 }) => {
+  const [previevMovie, setPrevievMovie] = useState<Movie | null>(null);
+  const [query, setQuery] = useState('');
+  const [noMovieFound, setNoMovieFound] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
+  const [badRequest, setBadRequest] = useState(false);
+
+  const isResponseError = useCallback((
+    response: MovieData | ResponseError,
+  ): response is ResponseError => {
+    return (response as ResponseError).Response === 'False';
+  }, []);
+
+  const loadMovie = useCallback(async (searchItem: string) => {
+    const foundMovie = await getMovie(searchItem);
+
+    if (isResponseError(foundMovie)) {
+      setNoMovieFound(true);
+      setIsLoading(false);
+
+      return;
+    }
+
+    const {
+      Title,
+      Plot,
+      imdbID,
+      Poster,
+    } = foundMovie as MovieData;
+
+    setPrevievMovie({
+      title: Title,
+      description: Plot,
+      imdbId: imdbID,
+      imgUrl: Poster
+        || 'https://via.placeholder.com/360x270.png?text=no%20preview',
+      imdbUrl: `https://www.imdb.com/title/${imdbID}/`,
+    });
+    setNoMovieFound(false);
+    setIsLoading(false);
+  }, []);
+
+  const handleSearchByQuery = useCallback((searchItem: string) => {
+    loadMovie(searchItem);
+  }, [loadMovie]);
+
+  const handleSetQuery = (event:React.ChangeEvent<HTMLInputElement>) => {
+    setQuery(event.target.value);
+  };
+
+  const handleAddToList = () => {
+    if (!movies.find(movie => movie.imdbId === previevMovie?.imdbId)) {
+      onAddToList(previevMovie as Movie);
+    }
+
+    setPrevievMovie(null);
+    setQuery('');
+  };
+
+  const handleLoading = () => {
+    setIsLoading(true);
+  };
+
+  if (query && query.trim().length === 0) {
+    setBadRequest(true);
+  } else {
+    setBadRequest(false);
+  }
+
   return (
     <>
       <form
         className="find-movie"
         onSubmit={(event) => {
-          onSubmit(query);
+          handleSearchByQuery(query);
           event.preventDefault();
-          onLoading();
+          handleLoading();
         }}
       >
         <div className="field">
@@ -50,11 +107,11 @@ export const FindMovie: React.FC<Props> = ({
               placeholder="Enter a title to search"
               className="input is-dander"
               value={query}
-              onChange={(event) => onChangeQuery(event)}
+              onChange={handleSetQuery}
             />
           </div>
 
-          {(noMovieFound || badRequest ) && (
+          {(noMovieFound || badRequest) && (
             <p className="help is-danger" data-cy="errorMessage">
               {badRequest
                 ? 'Put the right movie name'
@@ -84,7 +141,7 @@ export const FindMovie: React.FC<Props> = ({
                 data-cy="addButton"
                 type="button"
                 className="button is-primary"
-                onClick={onAddToList}
+                onClick={handleAddToList}
               >
                 Add to the list
               </button>
